--- src/django_livefield/__init__.py
@@ -1,19 +1,0 @@
-from .live_field import LiveField  # noqa
-from .live_manager import LiveManager  # noqa
-from .live_model import LiveModel  # noqa
-from .live_queryset import LiveQuerySet  # noqa
-
-
-__all__ = (
-    'LiveField',
-    'LiveManager',
-    'LiveModel',
-    'LiveQuerySet',
-)
-
-
-__version__ = 'unknown'
-try:
-    __version__ = __import__('pkg_resources').get_distribution('django_livefield').version
-except Exception as e:
-    pass
